requirements.txt
python-dotenv==0.10.1
flask==1.0.3
SQLAlchemy==1.1.11
